LeggettPlus

LeggettPlus provides an open-source, reproducible implementation of the Leggett+ physiologically based pharmacokinetic (PBPK) model for lead.

Details

This package was created by the Office of Environmental Health Hazard Assessment (OEHHA), an office within the State of California's Environmental Protection Agency (CalEPA). It is intended for research and exposure-scenario exploration only. Outputs should be interpreted within the context of model assumptions, parameter uncertainty, and available toxicokinetic data. This package should not be used and is not intended to be used for clinical diagnosis or individual medical decision-making. OEHHA provides no warranty for this package, either express or implied, including but not limited to the implied warranties of merchantability or fitness for a particular purpose; it is provided "AS IS." OEHHA disclaims any liability resulting from the use of this package.

Physiological parameters (body weight, hematocrit, blood-volume scaling, and RBC binding capacity) are derived from adult male / male-dominated occupational-worker data (Leggett, 1993, and subsequent updates). There is no female-specific parameter set implemented (e.g. a lower hematocrit range or pregnancy-related blood-volume expansion); results should not be assumed to represent female physiology.

See also

Author

Maintainer: Scott Coffin scott.l.coffin@gmail.com (ORCID)

Authors:

Scott Coffin scott.l.coffin@gmail.com (ORCID)

Hsing-Chieh Lin hclin@tamu.edu

Other contributors:

New Toxicology Evaluations Section, Office of Environmental Health Hazard Assessment NTES@oehha.ca.gov [copyright holder, funder]